fix | heap use after free | KqueueLoop.cpp

+++ include/struct.hpp
@@ -5,8 +5,8 @@
 
 struct eventStatus {
   int status; // WRITE_MODE, READ_MODE, NEW_EVENT, PROCESSOR, READ_ERROR ...
-  uintptr_t ident;
-  uintptr_t sub;
+  int ident;
+  int sub;
 };
 
 struct eventInfo {
